requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile --refresh --all-extras pyproject.toml -o requirements.txt
altair==5.5.0
    # via streamlit
annotated-types==0.7.0
    # via pydantic
appnope==0.1.4
    # via ipykernel
asttokens==3.0.0
    # via stack-data
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
blinker==1.9.0
    # via streamlit
boto3==1.38.41
    # via e84-geoai-common
boto3-stubs==1.38.41
    # via e84-geoai-common
botocore==1.38.41
    # via
    #   boto3
    #   s3transfer
botocore-stubs==1.38.30
    # via boto3-stubs
branca==0.8.1
    # via
    #   folium
    #   streamlit-folium
build==1.2.2.post1
    # via natural-language-geocoding (pyproject.toml)
cachetools==6.1.0
    # via streamlit
certifi==2025.6.15
    # via
    #   opensearch-py
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via streamlit
colorama==0.4.6
    # via pytest-watch
comm==0.2.2
    # via ipykernel
debugpy==1.8.14
    # via ipykernel
decorator==5.2.1
    # via ipython
deepdiff==8.5.0
    # via natural-language-geocoding (pyproject.toml)
distlib==0.3.9
    # via virtualenv
docopt==0.6.2
    # via pytest-watch
e84-geoai-common @ git+https://github.com/Element84/e84-geoai-common@9698f9a86e7d27d1e808e71a6849d252244fa2fc
    # via natural-language-geocoding (pyproject.toml)
events==0.5
    # via opensearch-py
executing==2.2.0
    # via stack-data
filelock==3.18.0
    # via virtualenv
folium==0.20.0
    # via streamlit-folium
function-schema==0.4.5
    # via e84-geoai-common
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via streamlit
identify==2.6.12
    # via pre-commit
idna==3.10
    # via requests
iniconfig==2.1.0
    # via pytest
ipykernel==6.29.5
    # via natural-language-geocoding (pyproject.toml)
ipython==9.3.0
    # via ipykernel
ipython-pygments-lexers==1.1.1
    # via ipython
jedi==0.19.2
    # via ipython
jinja2==3.1.6
    # via
    #   altair
    #   branca
    #   folium
    #   pydeck
    #   streamlit-folium
jmespath==1.0.1
    # via
    #   boto3
    #   botocore
jsonschema==4.24.0
    # via altair
jsonschema-specifications==2025.4.1
    # via jsonschema
jupyter-client==8.6.3
    # via ipykernel
jupyter-core==5.8.1
    # via
    #   ipykernel
    #   jupyter-client
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib-inline==0.1.7
    # via
    #   ipykernel
    #   ipython
mdurl==0.1.2
    # via markdown-it-py
mypy-boto3-bedrock==1.38.41
    # via boto3-stubs
mypy-boto3-bedrock-runtime==1.38.4
    # via
    #   boto3-stubs
    #   e84-geoai-common
mypy-boto3-s3==1.38.39
    # via boto3-stubs
narwhals==1.43.1
    # via altair
nest-asyncio==1.6.0
    # via ipykernel
nodeenv==1.9.1
    # via
    #   pre-commit
    #   pyright
numpy==2.3.1
    # via
    #   folium
    #   pandas
    #   pydeck
    #   shapely
    #   streamlit
    #   types-shapely
opensearch-py==3.0.0
    # via natural-language-geocoding (pyproject.toml)
orderly-set==5.4.1
    # via deepdiff
packaging==25.0
    # via
    #   altair
    #   build
    #   ipykernel
    #   pytest
    #   streamlit
pandas==2.3.0
    # via streamlit
parso==0.8.4
    # via jedi
pexpect==4.9.0
    # via ipython
pillow==11.2.1
    # via streamlit
platformdirs==4.3.8
    # via
    #   jupyter-core
    #   virtualenv
pluggy==1.6.0
    # via pytest
pre-commit==4.2.0
    # via natural-language-geocoding (pyproject.toml)
pre-commit-hooks==5.0.0
    # via natural-language-geocoding (pyproject.toml)
prompt-toolkit==3.0.51
    # via ipython
protobuf==6.31.1
    # via streamlit
psutil==7.0.0
    # via ipykernel
ptyprocess==0.7.0
    # via pexpect
pure-eval==0.2.3
    # via stack-data
pyarrow==20.0.0
    # via streamlit
pydantic==2.11.7
    # via e84-geoai-common
pydantic-core==2.33.2
    # via pydantic
pydeck==0.9.1
    # via streamlit
pygments==2.19.2
    # via
    #   ipython
    #   ipython-pygments-lexers
    #   pytest
    #   rich
pyproject-hooks==1.2.0
    # via build
pyright==1.1.402
    # via natural-language-geocoding (pyproject.toml)
pytest==8.4.1
    # via
    #   natural-language-geocoding (pyproject.toml)
    #   pytest-watch
pytest-watch==4.2.0
    # via natural-language-geocoding (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   botocore
    #   jupyter-client
    #   opensearch-py
    #   pandas
pytz==2025.2
    # via pandas
pyyaml==6.0.2
    # via pre-commit
pyzmq==27.0.0
    # via
    #   ipykernel
    #   jupyter-client
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   natural-language-geocoding (pyproject.toml)
    #   folium
    #   opensearch-py
    #   streamlit
rich==14.0.0
    # via
    #   natural-language-geocoding (pyproject.toml)
    #   e84-geoai-common
rpds-py==0.25.1
    # via
    #   jsonschema
    #   referencing
ruamel-yaml==0.18.14
    # via pre-commit-hooks
ruamel-yaml-clib==0.2.12
    # via ruamel-yaml
ruff==0.12.0
    # via natural-language-geocoding (pyproject.toml)
s3transfer==0.13.0
    # via boto3
shapely==2.1.1
    # via e84-geoai-common
six==1.17.0
    # via
    #   python-dateutil
    #   zss
smmap==5.0.2
    # via gitdb
stack-data==0.6.3
    # via ipython
streamlit==1.46.0
    # via
    #   natural-language-geocoding (pyproject.toml)
    #   streamlit-folium
streamlit-folium==0.25.0
    # via natural-language-geocoding (pyproject.toml)
tabulate==0.9.0
    # via natural-language-geocoding (pyproject.toml)
tenacity==9.1.2
    # via streamlit
toml==0.10.2
    # via streamlit
tornado==6.5.1
    # via
    #   ipykernel
    #   jupyter-client
    #   streamlit
traitlets==5.14.3
    # via
    #   comm
    #   ipykernel
    #   ipython
    #   jupyter-client
    #   jupyter-core
    #   matplotlib-inline
types-awscrt==0.27.2
    # via botocore-stubs
types-s3transfer==0.13.0
    # via boto3-stubs
types-shapely==2.1.0.20250512
    # via e84-geoai-common
typing-extensions==4.14.0
    # via
    #   altair
    #   pydantic
    #   pydantic-core
    #   pyright
    #   referencing
    #   streamlit
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
tzdata==2025.2
    # via pandas
urllib3==2.5.0
    # via
    #   botocore
    #   opensearch-py
    #   requests
virtualenv==20.31.2
    # via pre-commit
watchdog==6.0.0
    # via
    #   natural-language-geocoding (pyproject.toml)
    #   pytest-watch
wcwidth==0.2.13
    # via prompt-toolkit
xyzservices==2025.4.0
    # via folium
zss==1.2.0
    # via natural-language-geocoding (pyproject.toml)
